backend/requirements.txt
# --- API & utils ---
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# --- TensorFlow track (Apple Silicon) ---
tensorflow==2.17.1
keras-cv==0.8.2
tf-keras==2.17.0          # shim agar Transformers 4.35 jalan dgn Keras 3

# --- HF stack (matching Transformers 4.35) ---
transformers==4.35.0
tokenizers==0.14.1
huggingface_hub==0.17.3   # penting: 0.23.x bikin konflik dgn tokenizers 0.14.x

# --- Vector DB & tools ---
chromadb==0.4.18
scikit-learn==1.4.2
Pillow==10.1.0

# --- Pydantic / protobuf pin buat kompat --
pydantic==2.5.0
protobuf>=3.20.3,<5
